requirements.txt
Flask>=2.2.5,<3.0
Flask-SQLAlchemy>=3.0.3,<4.0
Flask-Migrate>=4.0.4,<5.0
Flask-Cors>=3.0.10,<4.0
Flask-JWT-Extended>=4.4.4,<5.0
Flask-Bcrypt>=1.0.1,<2.0
Flask-Mail>=0.9.1,<1.0
python-dotenv>=1.0.0,<2.0
psycopg2-binary>=2.9.5,<3.0
gunicorn==20.1.0
Werkzeug>=2.2.3,<3.0
SQLAlchemy>=2.0.4,<3.0
alembic>=1.9.4,<2.0
PyJWT>=2.6.0,<3.0
bcrypt>=4.0.1,<5.0
blinker>=1.5,<2.0
itsdangerous>=2.1.2,<3.0
Jinja2>=3.1.2,<4.0
MarkupSafe>=2.1.2,<3.0